+++ app/src/main/java/woowacourse/shopping/data/db/product/ProductService.kt
@@ -10,5 +10,5 @@ interface ProductService {
         offset: Int,
     ): List<Product>
 
-    fun findProductById(id: Long): Product
+    fun findProductById(id: Int): Product
 }
