+++ app/src/main/java/woowacourse/shopping/domain/model/CartItem.kt
@@ -1,12 +1,12 @@
 package woowacourse.shopping.domain.model
 
-import woowacourse.shopping.data.db.cart.CartRepository.Companion.DEFAULT_QUANTITY
+import woowacourse.shopping.domain.repository.CartRepository.Companion.DEFAULT_QUANTITY
 
 data class CartItem(
-    val id: Long,
-    val productId: Long,
+    val id: Int,
+    val productId: Int,
     val productName: String,
-    val price: Long,
+    val price: Int,
     val imageUrl: String,
     val quantity: Int,
 ) {
@@ -18,5 +18,5 @@ data class CartItem(
         return copy(quantity = (this.quantity - 1).coerceAtLeast(DEFAULT_QUANTITY))
     }
 
-    fun totalPrice(): Long = price * quantity
+    fun totalPrice(): Int = price * quantity
 }
