+++ app/src/main/java/woowacourse/shopping/view/detail/DetailViewModelFactory.kt
@@ -2,15 +2,15 @@ package woowacourse.shopping.view.detail
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import woowacourse.shopping.data.db.cart.CartRepository
-import woowacourse.shopping.data.db.product.ProductRepository
-import woowacourse.shopping.data.db.recent.RecentProductRepository
+import woowacourse.shopping.domain.repository.CartRepository
+import woowacourse.shopping.domain.repository.ProductRepository
+import woowacourse.shopping.domain.repository.RecentProductRepository
 
 class DetailViewModelFactory(
     private val cartRepository: CartRepository,
     private val productRepository: ProductRepository,
     private val recentProductRepository: RecentProductRepository,
-    private val productId: Long,
+    private val productId: Int,
 ) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         if (modelClass.isAssignableFrom(DetailViewModel::class.java)) {
