+++ app/src/test/java/woowacourse/shopping/helper/GetOrAwaitValue.kt
@@ -7,7 +7,7 @@ import java.util.concurrent.TimeUnit
 import java.util.concurrent.TimeoutException
 
 fun <T> LiveData<T>.getOrAwaitValue(
-    time: Long = 2,
+    time: Int = 2,
     timeUnit: TimeUnit = TimeUnit.SECONDS,
 ): T {
     var data: T? = null
